+++ packages/generator/src/index.ts
@@ -1,2 +1,3 @@
 export * from './readme.js';
 export * from './site-data.js';
+export * from './recommend.js';